Move SQL lookup to Persistence
