+++ src/WikibaseRdfExtension.php
@@ -6,8 +6,7 @@ namespace ProfessionalWiki\WikibaseRDF;
 
 use MediaWiki\MediaWikiServices;
 use MediaWiki\Rest\ResponseFactory;
-use ProfessionalWiki\WikibaseRDF\Application\GetAllMappings\AllMappingsLookup;
-use ProfessionalWiki\WikibaseRDF\Application\GetAllMappings\SqlAllMappingsLookup;
+use ProfessionalWiki\WikibaseRDF\Application\AllMappingsLookup;
 use ProfessionalWiki\WikibaseRDF\Application\MappingRepository;
 use ProfessionalWiki\WikibaseRDF\Application\SaveMappings\SaveMappingsPresenter;
 use ProfessionalWiki\WikibaseRDF\Application\SaveMappings\SaveMappingsUseCase;
@@ -16,6 +15,7 @@ use ProfessionalWiki\WikibaseRDF\EntryPoints\Rest\GetAllMappingsApi;
 use ProfessionalWiki\WikibaseRDF\Persistence\ContentSlotMappingRepository;
 use ProfessionalWiki\WikibaseRDF\Persistence\EntityContentRepository;
 use ProfessionalWiki\WikibaseRDF\Persistence\SlotEntityContentRepository;
+use ProfessionalWiki\WikibaseRDF\Persistence\SqlAllMappingsLookup;
 use ProfessionalWiki\WikibaseRDF\Presentation\MappingsPresenter;
 use ProfessionalWiki\WikibaseRDF\Presentation\RDF\MappingRdfBuilder;
 use ProfessionalWiki\WikibaseRDF\EntryPoints\Rest\GetMappingsApi;
